chore: create dedicated file for WebMessageReceived Event
