+++ pkg/edge/ICoreWebView2WebMessageReceivedEventArgs.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package edge
+
+type iCoreWebView2WebMessageReceivedEventArgsVtbl struct {
+	_IUnknownVtbl
+	GetSource                ComProc
+	GetWebMessageAsJSON      ComProc
+	TryGetWebMessageAsString ComProc
+}
+
+type iCoreWebView2WebMessageReceivedEventArgs struct {
+	vtbl *iCoreWebView2WebMessageReceivedEventArgsVtbl
+}
