+++ pkg/edge/corewebview2.go
@@ -272,19 +272,6 @@ func (e *ICoreWebView2Environment) CreateWebResourceResponse(content []byte, sta
 
 }
 
-// ICoreWebView2WebMessageReceivedEventArgs
-
-type iCoreWebView2WebMessageReceivedEventArgsVtbl struct {
-	_IUnknownVtbl
-	GetSource                ComProc
-	GetWebMessageAsJSON      ComProc
-	TryGetWebMessageAsString ComProc
-}
-
-type iCoreWebView2WebMessageReceivedEventArgs struct {
-	vtbl *iCoreWebView2WebMessageReceivedEventArgsVtbl
-}
-
 // ICoreWebView2PermissionRequestedEventArgs
 
 type iCoreWebView2PermissionRequestedEventArgsVtbl struct {
@@ -346,55 +333,6 @@ var iCoreWebView2CreateCoreWebView2EnvironmentCompletedHandlerFn = iCoreWebView2
 func newICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler(impl iCoreWebView2CreateCoreWebView2EnvironmentCompletedHandlerImpl) *iCoreWebView2CreateCoreWebView2EnvironmentCompletedHandler {
 	return &iCoreWebView2CreateCoreWebView2EnvironmentCompletedHandler{
 		vtbl: &iCoreWebView2CreateCoreWebView2EnvironmentCompletedHandlerFn,
-		impl: impl,
-	}
-}
-
-// ICoreWebView2WebMessageReceivedEventHandler
-
-type iCoreWebView2WebMessageReceivedEventHandlerImpl interface {
-	_IUnknownImpl
-	MessageReceived(sender *ICoreWebView2, args *iCoreWebView2WebMessageReceivedEventArgs) uintptr
-}
-
-type iCoreWebView2WebMessageReceivedEventHandlerVtbl struct {
-	_IUnknownVtbl
-	Invoke ComProc
-}
-
-type iCoreWebView2WebMessageReceivedEventHandler struct {
-	vtbl *iCoreWebView2WebMessageReceivedEventHandlerVtbl
-	impl iCoreWebView2WebMessageReceivedEventHandlerImpl
-}
-
-func _ICoreWebView2WebMessageReceivedEventHandlerIUnknownQueryInterface(this *iCoreWebView2WebMessageReceivedEventHandler, refiid, object uintptr) uintptr {
-	return this.impl.QueryInterface(refiid, object)
-}
-
-func _ICoreWebView2WebMessageReceivedEventHandlerIUnknownAddRef(this *iCoreWebView2WebMessageReceivedEventHandler) uintptr {
-	return this.impl.AddRef()
-}
-
-func _ICoreWebView2WebMessageReceivedEventHandlerIUnknownRelease(this *iCoreWebView2WebMessageReceivedEventHandler) uintptr {
-	return this.impl.Release()
-}
-
-func _ICoreWebView2WebMessageReceivedEventHandlerInvoke(this *iCoreWebView2WebMessageReceivedEventHandler, sender *ICoreWebView2, args *iCoreWebView2WebMessageReceivedEventArgs) uintptr {
-	return this.impl.MessageReceived(sender, args)
-}
-
-var iCoreWebView2WebMessageReceivedEventHandlerFn = iCoreWebView2WebMessageReceivedEventHandlerVtbl{
-	_IUnknownVtbl{
-		NewComProc(_ICoreWebView2WebMessageReceivedEventHandlerIUnknownQueryInterface),
-		NewComProc(_ICoreWebView2WebMessageReceivedEventHandlerIUnknownAddRef),
-		NewComProc(_ICoreWebView2WebMessageReceivedEventHandlerIUnknownRelease),
-	},
-	NewComProc(_ICoreWebView2WebMessageReceivedEventHandlerInvoke),
-}
-
-func newICoreWebView2WebMessageReceivedEventHandler(impl iCoreWebView2WebMessageReceivedEventHandlerImpl) *iCoreWebView2WebMessageReceivedEventHandler {
-	return &iCoreWebView2WebMessageReceivedEventHandler{
-		vtbl: &iCoreWebView2WebMessageReceivedEventHandlerFn,
 		impl: impl,
 	}
 }
